Move *Event classes into .parser.event package
Refactoring: Move Class
Target: *Event
Reason: To collect relational classes at same package
