--- src/com/esotericsoftware/yamlbeans/parser/EventType.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2008 Nathan Sweet
- * 
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software
- * is furnished to do so, subject to the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
- * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
- * IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package com.esotericsoftware.yamlbeans.parser;
-
-/** @author <a href="mailto:misc@n4te.com">Nathan Sweet</a> */
-public enum EventType {
-	STREAM_START, //
-	STREAM_END, //
-	SEQUENCE_START, //
-	SEQUENCE_END, //
-	SCALAR, //
-	MAPPING_START, //
-	MAPPING_END, //
-	DOCUMENT_START, //
-	DOCUMENT_END, //
-	ALIAS;
-
-	public String toString () {
-		return name().toLowerCase().replace('_', ' ');
-	}
-}
