+++ src/com/esotericsoftware/yamlbeans/parser/Production.java
@@ -1,5 +1,7 @@
 package com.esotericsoftware.yamlbeans.parser;
 
+import com.esotericsoftware.yamlbeans.parser.event.Event;
+
 public abstract class Production {
 	public abstract Event produce ();
 }
