+++ src/com/esotericsoftware/yamlbeans/parser/event/SequenceStartEvent.java
@@ -14,28 +14,12 @@
  * IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.esotericsoftware.yamlbeans.parser;
-
-import java.util.Arrays;
+package com.esotericsoftware.yamlbeans.parser.event;
 
 /** @author <a href="mailto:misc@n4te.com">Nathan Sweet</a>
  * @author <a href="mailto:ola.bini@ki.se">Ola Bini</a> */
-public class ScalarEvent extends NodeEvent {
-	public final String tag;
-	public final boolean[] implicit;
-	public final String value;
-	public final char style;
-
-	public ScalarEvent (String anchor, String tag, boolean[] implicit, String value, char style) {
-		super(EventType.SCALAR, anchor);
-		this.tag = tag;
-		this.implicit = implicit;
-		this.value = value;
-		this.style = style;
-	}
-
-	public String toString () {
-		return "<" + type + " value='" + value + "' anchor='" + anchor + "' tag='" + tag + "' implicit='"
-			+ Arrays.toString(implicit) + "' style='" + (style == 0 ? "" : style) + "'>";
+public class SequenceStartEvent extends CollectionStartEvent {
+	public SequenceStartEvent (String anchor, String tag, boolean implicit, boolean flowStyle) {
+		super(EventType.SEQUENCE_START, anchor, tag, implicit, flowStyle);
 	}
 }
